+++ server/db/index.tsx
@@ -1,11 +1,12 @@
 import { MongoClient } from "mongodb";
 import { Sequelize } from "sequelize-typescript";
 import { mongoConfig, psqlConfig } from "./configs";
-import SiteTheme from "./models_psql/site-theme";
-import UserTheme from "./models_psql/user-theme";
+import SiteTheme from "./models/psql/site-theme";
+import User from "./models/psql/user";
+import UserTheme from "./models/psql/user-theme";
 
 const psqlConnection = new Sequelize(psqlConfig);
-psqlConnection.addModels([UserTheme, SiteTheme]);
+psqlConnection.addModels([UserTheme, SiteTheme, User]);
 
 const mongoConnection = new MongoClient(mongoConfig);
